master brach alignement (#135)
* PHP7 Support badge remove

PHP7 is supported by default so this badge is now useless.

* Revert Yaml to PHP (#116)

* Minor tests modifications

* Jwt split (#118)

* JWS/JWE Split

THis PR adds a `split` method for both JWS and JWE objects.
It allows to explode a JWS/JWE that contains several signatures/recipients into a list of JWS/JWE with only one signature/recipient.

It can be useful in certain cases e.g. where the verification of each signature is required or if a JWE has to be computed at once, but sent separetly to recipients.

* Keys in a key set is a list, not an object with indexes

* Deprecation fixed

* Bug fixed

Routes with :: does not work properly with Symfony 3.
x if he controller is a service

* Bug fixed (#132)

EC keys were not correctly created using the OpenSSL method.